+++ commands/actions.go
@@ -8,11 +8,29 @@ import (
 	"github.com/leanovate/gopter/gen"
 )
 
+type shrinkableCommand struct {
+	command  Command
+	shrinker gopter.Shrinker
+}
+
+func (s shrinkableCommand) shrink() gopter.Shrink {
+	return s.shrinker(s.command).Map(func(command Command) shrinkableCommand {
+		return shrinkableCommand{
+			command:  command,
+			shrinker: s.shrinker,
+		}
+	})
+}
+
+func (s shrinkableCommand) String() string {
+	return fmt.Sprintf("%v", s.command)
+}
+
 type actions struct {
 	// initialStateProvider has to reset/recreate the initial state exactly the
 	// same every time.
 	initialStateProvider func() State
-	sequentialCommands   []Command
+	sequentialCommands   []shrinkableCommand
 	// parallel commands will come later
 }
 
@@ -23,25 +41,28 @@ func (a *actions) String() string {
 func (a *actions) run(systemUnderTest SystemUnderTest) (*gopter.PropResult, error) {
 	state := a.initialStateProvider()
 	propResult := &gopter.PropResult{Status: gopter.PropTrue}
-	for _, command := range a.sequentialCommands {
-		if !command.PreCondition(state) {
+	for _, shrinkableCommand := range a.sequentialCommands {
+		if !shrinkableCommand.command.PreCondition(state) {
 			return &gopter.PropResult{Status: gopter.PropFalse}, nil
 		}
-		result := command.Run(systemUnderTest)
-		state = command.NextState(state)
-		propResult = propResult.And(command.PostCondition(state, result))
+		result := shrinkableCommand.command.Run(systemUnderTest)
+		state = shrinkableCommand.command.NextState(state)
+		propResult = propResult.And(shrinkableCommand.command.PostCondition(state, result))
 	}
 	return propResult, nil
 }
 
 type sizedCommands struct {
 	state    State
-	commands []Command
+	commands []shrinkableCommand
 }
 
 func actionsShrinker(v interface{}) gopter.Shrink {
 	a := v.(*actions)
-	return gen.SliceShrinker(gopter.NoShrinker)(a.sequentialCommands).Map(func(v []Command) *actions {
+	elementShrinker := gopter.Shrinker(func(v interface{}) gopter.Shrink {
+		return v.(shrinkableCommand).shrink()
+	})
+	return gen.SliceShrinker(elementShrinker)(a.sequentialCommands).Map(func(v []shrinkableCommand) *actions {
 		return &actions{
 			initialStateProvider: a.initialStateProvider,
 			sequentialCommands:   v,
@@ -73,11 +94,11 @@ func genActions(commands Commands) gopter.Gen {
 			}
 		}).SuchThat(func(actions *actions) bool {
 			state := actions.initialStateProvider()
-			for _, command := range actions.sequentialCommands {
-				if !command.PreCondition(state) {
+			for _, shrinkableCommand := range actions.sequentialCommands {
+				if !shrinkableCommand.command.PreCondition(state) {
 					return false
 				}
-				state = command.NextState(state)
+				state = shrinkableCommand.command.NextState(state)
 			}
 			return true
 		}).WithShrinker(actionsShrinker)
@@ -88,18 +109,29 @@ func genSizedCommands(commands Commands, initialStateProvider func() State) gopt
 	return func(genParams *gopter.GenParameters) *gopter.GenResult {
 		sizedCommandsGen := gen.Const(sizedCommands{
 			state:    initialStateProvider(),
-			commands: make([]Command, 0, genParams.Size),
+			commands: make([]shrinkableCommand, 0, genParams.Size),
 		})
 		for i := 0; i < genParams.Size; i++ {
 			sizedCommandsGen = sizedCommandsGen.FlatMap(func(v interface{}) gopter.Gen {
 				prev := v.(sizedCommands)
 				return gen.RetryUntil(commands.GenCommand(prev.state), func(command Command) bool {
 					return command.PreCondition(prev.state)
-				}, 100).Map(func(command Command) sizedCommands {
-					return sizedCommands{
-						state:    command.NextState(prev.state),
-						commands: append(prev.commands, command),
+				}, 100).MapResult(func(result *gopter.GenResult) *gopter.GenResult {
+					value, ok := result.Retrieve()
+					if !ok {
+						return gopter.NewEmptyResult(reflect.TypeOf(sizedCommands{}))
 					}
+					command := value.(Command)
+					return gopter.NewGenResult(
+						sizedCommands{
+							state: command.NextState(prev.state),
+							commands: append(prev.commands, shrinkableCommand{
+								command:  command,
+								shrinker: result.Shrinker,
+							}),
+						},
+						gopter.NoShrinker,
+					)
 				})
 			}, reflect.TypeOf(sizedCommands{}))
 		}