+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Additions
 - `gopter.GenParameters` now has a `CloneWithSeed(seed int64)` function to
   temparary copies to create rerunable sections of code.
+- Added `gopter.Gen.MapResult` for power-user mappings
 
 ### Changed
 - Refactored `commands` package under the hood to allow the use of mutable state.